음료요청게시판 상세조회 기능 추가

+++ src/main/java/com/springboot/board/model/CafeRequest.java
@@ -34,7 +34,7 @@ public class CafeRequest {
 			)
 	
 	@Column(name = "BNO")
-	private Long bno;
+	private Integer bno;
 	
 	@Column(name = "TITLE")
 	private String title;
